🐛 fix : 스크랩 취소 API 500 에러 수정

+++ src/main/java/com/inuteamflow/server/domain/infoPost/service/InfoPostService.java
@@ -245,7 +245,7 @@ public class InfoPostService {
     public void unscrapInfoPost(Long infoPostId, User user) {
         InfoPost infoPost = getInfoPostById(infoPostId);
 
-        long deleted = infoPostScrapRepository.deleteByInfoPostAndUser(infoPost, user);
+        int deleted = infoPostScrapRepository.deleteByInfoPostAndUser(infoPost, user);
         if (deleted == 0) {
             throw new RestApiException(CustomErrorCode.INFO_POST_SCRAP_NOT_FOUND);
         }
